OCPBUGS-26498: Deprecation Validation plugin
Add deprecation validation plugin which provides a framework for adding
a deprecated status to a route. Update the RouteRejectionRecorder
interface to be the RouteStatusRecorder to support more status-related
use cases.

For this initial implementation, if a route has a SHA1 certificate,
the deprecation validation plugin will mark the route deprecated.