Add Upgrade Validation router plugin
Add the Upgrade Validation plugin which provides a framework for adding
a status to a route to indicate future upgrade issues. For this
initial implementation, if a route has a SHA1 certificate, the upgrade
validation plugin will mark the route with a UnservableInFutureVersions
status. If the route no longer has a SHA1 certificate, remove the status
instead of setting UnservableInFutureVersions=false.

This is specific to OpenShift 4.15 and is expected to be removed
from 4.16 once backported to avoid confusion. It's specific to 4.15
because 4.15 accepts SHA1 certs, but 4.16 does not. Therefore,
validating the upgrade should be only done in 4.15.

Update the RejectionRecorder interface to be the RouteStatusRecorder to
support more generic status-related use cases as we have more than just
Admitted status now.

Add UPGRADE_VALIDATION env variable, but default to true, so that we can
disable in the future. Follow existing pattern similar to existing
EXTENDED_VALIDATION env variable.

Key the writerlease logic off of route UID AND condition type. This is
now essential because if only route UID, then the lease logic will only
update one condition as it doesn't differentiate from the work items for
the different conditions.

Adjust findMostRecentIngress to use the latest condition instead of just
Admitted as we can no longer assume that Admitted is the only condition.

Add a type for writelease WorkKey to ensure strings aren't used anymore
as the previous logic was only using route.UID.