Fix recordIngressCondition changed=false issue.
When an route ingress already exists, but an expected condition isn't
found, recordIngressCondition would previously not change/update
anything. This wasn't an issue for when only had Admitted because the
route ingress was updated synchronously with the Admitted condition.
However, adding another route status condition exposes the scenario in
which a route ingress will exist, but the desired condition won't.

This also adds a unit test for recordIngressCondition which explicitly
tests this case along with other edge cases.