Remove original condition clearing from recordIngressCondition
Previously, recordIngressCondition sometimes clears the original return value
conditions depending on whether the condition was found and changed. This
doesn't seem consistent and makes it hard to unit test. Futhermore, the
original return value doesn't impact any production code logic and
appears to be added to support unit testing. Therefore, it seems reasonable
to remove condition clearing as there is no clear reason it is needed,
it is inconsistent, and confusing as the name "original" implies it
should be...original.

Also, make original use a deep copy. Though shallow copy works for now,
it's dangerous for future developers since it's not actually making a
copy.

Update unit testing to now verify original return value from
recordIngressCondition.